refactor for app and data_store split
