+++ .github/workflows/build-test.yaml
@@ -1,0 +1,28 @@
+name: Code Check
+
+on:
+  push:
+    branches:
+      - development
+  pull_request:
+    branches:
+      - development
+
+jobs:
+  pre-commit:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          fetch-depth: 0 # Important for pre-commit to work correctly with staged changes
+
+      - uses: actions/setup-python@v4
+        with:
+          python-version: "3.12" # Match your project's Python version
+          cache: "pip"
+          cache-dependency-path: "pyproject.toml"
+
+      - name: Install dependencies
+        run: pip install -e .[dev]
+
+      - uses: pre-commit/action@v3.0.0
