+++ .pre-commit-config.yaml
@@ -1,106 +1,74 @@
 # See https://pre-commit.com for more information
 # See https://pre-commit.com/hooks.html for more hooks
 repos:
-  - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v5.0.0
-    hooks:
-      # Fails if there are any ">>>>>" lines in files due to merge conflicts.
-      - id: check-merge-conflict
-      # Trims trailing whitespace. Allow a single space on the end of .md lines for hard line breaks.
-      - id: trailing-whitespace
-        args: [--markdown-linebreak-ext=md]
-      # Makes sure files end in a newline and only a newline;
-      - id: end-of-file-fixer
-      # Attempts to load all TOML files to verify syntax.
-      - id: check-toml
-      # Attempts to load all yaml files to verify syntax; unsafe: only check syntax, do not load yaml
-      - id: check-yaml
-        args: ["--unsafe"]
-      # Check for symlinks that do not point to anything.
-      - id: check-symlinks
-      # Fail if staged files are above a certain size.
-      # To add a large file, use 'git lfs track <file>; git add <file> to track large files with
-      # git-lfs rather than commiting them directly to the git history
-      - id: check-added-large-files
-        args: ["--maxkb=500"]
+- repo: https://github.com/pre-commit/pre-commit-hooks
+  rev: v5.0.0
+  hooks:
+  # Fails if there are any ">>>>>" lines in files due to merge conflicts.
+  - id: check-merge-conflict
+  # Trims trailing whitespace. Allow a single space on the end of .md lines for hard line breaks.
+  - id: trailing-whitespace
+    args: [ --markdown-linebreak-ext=md ]
+  # Makes sure files end in a newline and only a newline;
+  - id: end-of-file-fixer
+  # Attempts to load all TOML files to verify syntax.
+  - id: check-toml
+  # Attempts to load all yaml files to verify syntax; unsafe: only check syntax, do not load yaml
+  - id: check-yaml
+    args: [ "--unsafe" ]
+  # Check for symlinks that do not point to anything.
+  - id: check-symlinks
+  # Fail if staged files are above a certain size.
+  # To add a large file, use 'git lfs track <file>; git add <file> to track large files with
+  # git-lfs rather than commiting them directly to the git history
+  - id: check-added-large-files
+    args: [ "--maxkb=500" ]
 
-        # HALT! Before you exclude a large file and commit it, forever
-        # bloating our repo size, did you:
-        # (1) use a CLI tool like imageoptim to compress them if they are images
-        # (2) think hard about whether using DVC or git-lfs is more appropriate
-        #     for the file--such as in the case of CSV files or other data
-        # This can be confusing. Reach out for help in our chat to help decide
-        # how to deal adding these large files you have :)
-        exclude: |
-          (?x)(
-              ^example/large/file.csv|
-              ^example/large/sklearn-model.pkl
-          )
-      # Sort requirements in requirements.txt files.
-      - id: requirements-txt-fixer
-      # Prevent addition of new git submodules.
-      - id: forbid-new-submodules
-      # Prevent committing directly to trunk (since Bitbucket wants us to pay for this feature)
-      - id: no-commit-to-branch
-        args: ["--branch=main"]
-      # Detects the presence of private keys
-      - id: detect-private-key
+    # HALT! Before you exclude a large file and commit it, forever
+    # bloating our repo size, did you:
+    # (1) use a CLI tool like imageoptim to compress them if they are images
+    # (2) think hard about whether using DVC or git-lfs is more appropriate
+    #     for the file--such as in the case of CSV files or other data
+    # This can be confusing. Reach out for help in our chat to help decide
+    # how to deal adding these large files you have :)
+    exclude: |
+      (?x)(
+          ^example/large/file.csv|
+          ^example/large/sklearn-model.pkl
+      )
+  # Sort requirements in requirements.txt files.
+  - id: requirements-txt-fixer
+  # Prevent addition of new git submodules.
+  - id: forbid-new-submodules
+  # Prevent committing directly to trunk (since Bitbucket wants us to pay for this feature)
+  - id: no-commit-to-branch
+    args: [ "--branch=main" ]
+  # Detects the presence of private keys
+  - id: detect-private-key
 
-  - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: "v0.9.6"
-    hooks:
-      - id: ruff
-        args: [--exit-non-zero-on-fix, --config=./pyproject.toml]
+- repo: https://github.com/pre-commit/mirrors-mypy
+  rev: "v1.15.0"
+  hooks:
+  - id: mypy
+    args: [ --no-strict-optional, --ignore-missing-imports, --config-file=./pyproject.toml ]
 
-  - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: "v1.15.0"
-    hooks:
-      - id: mypy
-        args:
-          [
-            --no-strict-optional,
-            --ignore-missing-imports,
-            --config-file=./pyproject.toml,
-          ]
+- repo: https://github.com/psf/black
+  rev: 25.1.0
+  hooks:
+  - id: black
+    args:
+    - --config=./pyproject.toml
 
-  - repo: https://github.com/psf/black
-    rev: 25.1.0
-    hooks:
-      - id: black
-        args:
-          - --config=./pyproject.toml
+- repo: https://github.com/PyCQA/pylint
+  rev: v3.3.4
+  hooks:
+  - id: pylint
+    args:
+    - --rcfile=./pyproject.toml
 
-  - repo: https://github.com/PyCQA/pylint
-    rev: v3.3.4
-    hooks:
-      - id: pylint
-        args:
-          - --rcfile=./pyproject.toml
-
-  - repo: https://github.com/PyCQA/flake8
-    rev: 7.1.1
-    hooks:
-      - id: flake8
-        args:
-          - --toml-config=./pyproject.toml
-        additional_dependencies:
-          - radon
-          - flake8-docstrings
-          - Flake8-pyproject
-
-  - repo: https://github.com/pycqa/isort
-    rev: 6.0.0
-    hooks:
-      - id: isort
-        args:
-          - --settings-path=./pyproject.toml
-
-  - repo: https://github.com/PyCQA/autoflake
-    rev: v2.3.1
-    hooks:
-      - id: autoflake
-        args:
-          - --in-place
-          - --remove-all-unused-imports
-          - --remove-unused-variable
-          - --ignore-init-module-imports
+- repo: https://github.com/pycqa/isort
+  rev: 6.0.0
+  hooks:
+  - id: isort
+    args:
+    - --settings-path=./pyproject.toml
