+++ pyproject.toml
@@ -1,33 +1,37 @@
 [project]
 name = "MG_questions_generator"
-authors = [{ name = "Ulises Jimenez", email = "ulisesjimenez@google.com" }, 
+authors = [{ name = "Ulises Jimenez", email = "ulisesjimenez@google.com" },
 { name = "Andres Toledo", email = "aatoledo@google.com" }]
 description = "Solution to create questions for a Meet and Greet"
 readme = "README.md"
 requires-python = ">=3.12"
 keywords = ["gcp", "genAI"]
-license = { text = "MIT" }
+license = { text = "Apache License" }
 classifiers = ["Programming Language :: Python :: 3"]
-dependencies = ["numpy", 'importlib-metadata; python_version<"3.8"']
+dependencies = ["gradio"]
 dynamic = ["version"]
 
+[project.optional-dependencies]
+dev = [
+    "pre-commit",
+    "mypy",
+    "black",
+    "pylint",
+    "isort",
+    "autoflake"
+]
 
 [tool.black]
 line-length = 119
-
-[tool.flake8]
-docstring-convention = "all"
-ignore = ["D107", "D212", "E501", "W503", "W605", "D203", "D100"]
-max-line-length = 119
-
-# radon
-radon-max-cc = 10
 
 [tool.isort]
 profile = "black"
 multi_line_output = "VERTICAL_HANGING_INDENT"
 force_grid_wrap = 2
 line_length = 119
+
+[tool.pylint.design]
+max-attributes = 10
 
 [tool.pylint."messages control"]
 disable = [
@@ -39,19 +43,3 @@ disable = [
     "too-few-public-methods",
     "redefined-outer-name",
 ]
-
-[tool.ruff]
-# Enable flake8-bugbear (`B`) rules.
-select = ["E", "F", "B", "ERA"]
-
-# Never enforce `E501` (line length violations).
-ignore = ["E501", "F401"]
-
-# Avoid trying to fix flake8-bugbear (`B`) violations.
-unfixable = ["B"]
-line-length = 119
-
-# Ignore `E402` (import violations) in all `__init__.py` files, and in `path/to/file.py`.
-[tool.ruff.per-file-ignores]
-"__init__.py" = ["E402"]
-"path/to/file.py" = ["E402"]
