+++ src/app/Dockerfile
@@ -1,0 +1,13 @@
+FROM python:3.12-slim
+
+RUN apt-get update
+
+WORKDIR /app
+COPY . .
+
+RUN pip install --upgrade pip
+RUN pip install -r requirements.txt
+
+EXPOSE 8080
+
+CMD [ "python",  "main.py"]
