+++ src/app/gemini_schema.py
@@ -9,27 +9,27 @@ response_schema = {
         "last_degree_studies",
         "soft_skills",
         "tech_stack",
-        "past_experience"
+        "past_experience",
     ],
     "properties": {
         "location": {
             "type": "string",
             "example": "Guadalajara, Jalisco, Mexico",
-            "description": "REQUIRED! The candidate location formatted as 'city, state, country'."
+            "description": "REQUIRED! The candidate location formatted as 'city, state, country'.",
         },
         "current_role": {
             "type": "string",
             "example": "Cloud AI Engineer",
-            "description": "REQUIRED! The current or latest role the candidate has held in a working environment."
+            "description": "REQUIRED! The current or latest role the candidate has held in a working environment.",
         },
         "current_company": {
             "type": "string",
             "example": "Google",
-            "description": "REQUIRED! The current or latest company the candidate has worked at."
+            "description": "REQUIRED! The current or latest company the candidate has worked at.",
         },
         "last_degree_studies": {
             "type": "string",
-            "description": "REQUIRED! The degree of the latest studies completed by the candidate, including area of study."
+            "description": "REQUIRED! The degree of the latest studies completed by the candidate, including area of study.",
         },
         "certifications": {
             "type": "array",
@@ -37,17 +37,17 @@ response_schema = {
             "items": {
                 "type": "string",
                 "example": "Professional Machine Learning Engineer, Google Cloud",
-                "description": "The certification held by the candidate and the company that validates the accreditation."
-            }
+                "description": "The certification held by the candidate and the company that validates the accreditation.",
+            },
         },
-        "soft_skills":{
+        "soft_skills": {
             "type": "array",
             "description": "List of candidate's soft skills.",
             "items": {
                 "type": "string",
                 "example": "team work",
-                "description": "A candidate's ability to work on their own or in a team."
-            }
+                "description": "A candidate's ability to work on their own or in a team.",
+            },
         },
         "tech_stack": {
             "type": "array",
@@ -55,19 +55,16 @@ response_schema = {
             "items": {
                 "type": "string",
                 "example": "Python",
-                "description": "A technologies (programming languages, frameworks, databases, tools, platforms, etc.) that is used to build and maintain applications or systems."
-            }
+                "description": "A technologies (programming languages, frameworks, databases, tools, platforms, etc.) that is used to build and maintain applications or systems.",
+            },
         },
         "past_experience": {
             "type": "array",
-            # Consider “pastExperience” as the relevant 
-            # past experience included in the resume. Use a dictionary structure to include the name of the company, 
-            # area, role, start and end dates, and a brief summary of the activities performed during each experience 
+            # Consider “pastExperience” as the relevant
+            # past experience included in the resume. Use a dictionary structure to include the name of the company,
+            # area, role, start and end dates, and a brief summary of the activities performed during each experience
             "description": "List of activities performed by the candidate during their last work experience. Provide at least 2.",
-            "items": {
-                "type": "string",
-                "description": "REQUIRED! Actions performed by the candidate."
-            }
-        }
-    }
+            "items": {"type": "string", "description": "REQUIRED! Actions performed by the candidate."},
+        },
+    },
 }
