+++ src/app/main.py
@@ -1,6 +1,8 @@
-import gradio as gr
+"""
+Gradio app for generating Meet & Greet questions based on a provided CV.
+"""
 
-# from google.cloud import storage
+import gradio as gr
 
 
 def get_base_file() -> gr.File:
@@ -13,12 +15,12 @@ def get_base_file() -> gr.File:
     )
 
 
-def get_base_markdown() -> None:
+def get_base_markdown():
     return gr.Markdown(
-        f"""
+        """
         ```
 
-        
+
 
 
 
@@ -27,17 +29,17 @@ def get_base_markdown() -> None:
     )
 
 
-def clean() -> None:
+def clean():
     return (
         gr.File(
             file_count="single",
             label="CV in PDF format",
         ),
         gr.Markdown(
-            f"""
+            """
         ```
 
-        
+
 
 
 
@@ -48,15 +50,14 @@ def clean() -> None:
 
 
 def summarize_file(cv_file) -> gr.Markdown:
-    # TODO:
     # storage_client = storage.Client()
     # bucket = storage_client.bucket(bucket_name)
     # blob = bucket.blob(destination_blob_name)
     # blob.upload_from_filename(file.name)
-
-    gr.Info(f"File processed successfully!")
+    print(cv_file)
+    gr.Info("File processed successfully!")
     return gr.Markdown(
-        f"""
+        """
                 It is a long established fact that a reader will be distracted by the readable content of a page when looking at its layout. The point of using Lorem Ipsum is that it has a more-or-less normal distribution of letters, as opposed to using 'Content here, content here', making it look like readable English. Many desktop publishing packages and web page editors now use Lorem Ipsum as their default model text, and a search for 'lorem ipsum' will uncover many web sites still in their infancy. Various versions have evolved over the years, sometimes by accident, sometimes on purpose (injected humour and the like).
                 """
     )
@@ -67,7 +68,7 @@ if __name__ == "__main__":
 
     with demo:
         gr.Markdown(
-            f"""
+            """
             # M&G Q-Gen
             Assistant for Meet & Greet question generation
             """
@@ -88,18 +89,13 @@ if __name__ == "__main__":
                     )
 
                 with gr.Row():
-                    btn_summarize = gr.Button(
-                        value="✨ Generate Questions!", interactive=True
-                    )
+                    btn_summarize = gr.Button(value="✨ Generate Questions!", interactive=True)
                     btn_clean = gr.Button(value="🗑️ Clean Up!", interactive=True)
 
             with gr.Column(visible=True) as summary_result:
                 summary_markdown = get_base_markdown()
-
-        btn_summarize.click(
-            fn=summarize_file, inputs=[cv_file], outputs=[summary_markdown]
-        )
-
+        # pylint: disable=E1101
+        btn_summarize.click(fn=summarize_file, inputs=[cv_file], outputs=[summary_markdown])
         btn_clean.click(fn=clean, inputs=[], outputs=[cv_file, summary_markdown])
 
     demo.launch(debug=True, server_name="0.0.0.0", server_port=8080)
