+++ src/app/requirements.txt
@@ -1,5 +1,5 @@
-gradio==5.15.0
 google-cloud-aiplatform==1.79.0
 google-cloud-bigquery==3.29.0
 google-cloud-storage==3.0.0
+gradio==5.15.0
 vertexai==1.71.1
