+++ src/data_store/requirements.txt
@@ -1,0 +1,1 @@
+google-cloud-discoveryengine
