--- src/requirements.txt
@@ -1,1 +1,0 @@
-google-cloud-discoveryengine
